Nodding Bug Patched
Fixed the nodding bug. The issue was caused by the yaw angle that it needed to be in and its actual. If the angle between them is to small the temi cannot move to get to the position. However, the temi is in the wrong spot. So the temi constantly tries to correct itself but cannot. There was a threshold to try and correct this, but it was not large enough. I increased it and it seems to have fixed it.